chore(layout): split three files to fix the red Rust Quality gate on main
`scripts/ci/check-openhuman-rust-layout.mjs` fails on pristine `main`
(run 33578839455 @ 61d25fe21), so `Rust Quality (fmt, clippy)` is red and
`PR CI Gate` blocks every open PR:

  - agent/harness/session/turn/context.rs: inline test module
  - tools/impl/filesystem/git_operations.rs: 949 lines (limit 750)
  - tools/impl/filesystem/git_operations_tests.rs: 929 lines (limit 750)

Layout only — no behaviour change, no test removed or ignored, no
`#[allow]` added, and the gate's limits are untouched.

Each split follows the convention already used elsewhere in the tree:

  - `context.rs` — the inline `#[cfg(test)] mod
    tool_policy_boundary_placement_tests` moves to a sibling
    `context_tests.rs`, declared as `#[cfg(test)] #[path = ...] mod tests;`
    (the form used at 872 other sites). All four tests move verbatim.

  - `git_operations.rs` — root keeps the `mod tests;` declaration and
    `include!`s `git_operations_part_01.rs` (435) + `_part_02.rs` (510),
    matching `flows/bus.rs` and `security/credentials/ops.rs`. The split
    falls on the closing brace of the first `impl GitOperationsTool`;
    concatenating the two parts reproduces the original body byte for byte.

  - `git_operations_tests.rs` — root keeps the imports and all six shared
    helpers, then declares `part_01_tests` / `part_02_tests` via `#[path]`,
    matching `flows/ops_tests.rs`. The helpers must sit in the root because
    both parts use them; each part opens with `use super::*;`. `#[cfg(unix)]`
    stays attached to `plant_fsmonitor_hook`, which it guards.

Verified: layout gate passes (0 violations), `cargo check -p openhuman
--tests` exits 0 with no new warnings, `cargo fmt --all -- --check` clean,
and the test-inventory guard reports 0 orphans. Test counts are unchanged —
42 `#[test]`/`#[tokio::test]` in the git_operations suite and 4 in context.
