+++ src/openhuman/tools/impl/filesystem/git_operations_part_01.rs
@@ -1,0 +1,435 @@
+use crate::openhuman::security::{AutonomyLevel, CommandClass, GateDecision, SecurityPolicy};
+use crate::openhuman::tools::traits::{Tool, ToolCallOptions, ToolResult};
+use async_trait::async_trait;
+use serde_json::json;
+use std::path::{Path, PathBuf};
+use std::sync::Arc;
+use tinytools::ToolRunContext;
+
+/// Git operations tool for structured repository management.
+/// Provides safe, parsed git operations with JSON output.
+pub struct GitOperationsTool {
+    security: Arc<SecurityPolicy>,
+    action_dir: PathBuf,
+}
+
+impl GitOperationsTool {
+    pub fn new(security: Arc<SecurityPolicy>, action_dir: PathBuf) -> Self {
+        Self {
+            security,
+            action_dir,
+        }
+    }
+
+    /// Resolve the working directory for git operations.
+    ///
+    /// Returns the per-worker git-worktree checkout when the tinyagents harness
+    /// threaded a [`WorkspaceDescriptor`] into this call's
+    /// [`ToolExecutionContext`] — an edit-capable worker running with
+    /// `isolation = "worktree"`, whose isolated worktree root is carried on the
+    /// run context (`RunContext::with_workspace`) and surfaced per tool call via
+    /// `ToolExecutionContext::from_run_context`. Otherwise falls back to the
+    /// tool's configured `action_dir`, which preserves the non-isolated
+    /// behaviour exactly. See #3376, #4249 (08.5).
+    fn effective_action_dir_for_context(&self, context: Option<&dyn ToolRunContext>) -> PathBuf {
+        if let Some(workspace) = context.and_then(|ctx| ctx.workspace()) {
+            tracing::debug!(
+                workspace_root = %workspace.root.display(),
+                policy_id = %workspace.policy_id,
+                "[git_operations] using TinyAgents workspace descriptor as action dir"
+            );
+            return workspace.root.clone();
+        }
+        self.action_dir.clone()
+    }
+
+    /// Sanitize git arguments to prevent injection attacks
+    fn sanitize_git_args(&self, args: &str) -> anyhow::Result<Vec<String>> {
+        let mut result = Vec::new();
+        for arg in args.split_whitespace() {
+            // Block dangerous git options that could lead to command injection
+            let arg_lower = arg.to_lowercase();
+            if arg_lower.starts_with("--exec=")
+                || arg_lower.starts_with("--upload-pack=")
+                || arg_lower.starts_with("--receive-pack=")
+                || arg_lower.starts_with("--pager=")
+                || arg_lower.starts_with("--editor=")
+                || arg_lower == "--no-verify"
+                || arg_lower.contains("$(")
+                || arg_lower.contains('`')
+                || arg.contains('|')
+                || arg.contains(';')
+                || arg.contains('>')
+            {
+                anyhow::bail!("Blocked potentially dangerous git argument: {arg}");
+            }
+            // Block `-c` config injection (exact match or `-c=...` prefix).
+            // This must not false-positive on `--cached` or `-cached`.
+            if arg_lower == "-c" || arg_lower.starts_with("-c=") {
+                anyhow::bail!("Blocked potentially dangerous git argument: {arg}");
+            }
+            result.push(arg.to_string());
+        }
+        Ok(result)
+    }
+
+    /// Check if an operation requires write access
+    fn requires_write_access(&self, operation: &str) -> bool {
+        matches!(
+            operation,
+            "commit" | "add" | "checkout" | "stash" | "reset" | "revert"
+        )
+    }
+
+    /// Check if an operation is read-only
+    fn is_read_only(&self, operation: &str) -> bool {
+        matches!(
+            operation,
+            "status" | "diff" | "log" | "show" | "branch" | "rev-parse"
+        )
+    }
+
+    async fn run_git_command_in(&self, cwd: &Path, args: &[&str]) -> anyhow::Result<String> {
+        if let Some(key) = first_disallowed_repo_config_key(cwd).await? {
+            tracing::debug!(
+                "[git_operations] refusing to run git: dir={}, disallowed_config_key={key}",
+                cwd.display()
+            );
+            anyhow::bail!(
+                "refusing to run git in {}: its repository config sets `{key}`, which is \
+                 not on the allowlist of configuration this tool will run under. \
+                 Several git config keys name a command git then executes, and this \
+                 directory is agent-writable, so unrecognised configuration is treated \
+                 as untrusted rather than honoured.",
+                cwd.display()
+            )
+        }
+
+        let output = hardened_git(cwd).args(args).output().await?;
+
+        if !output.status.success() {
+            let stderr = String::from_utf8_lossy(&output.stderr);
+            anyhow::bail!("Git command failed: {stderr}");
+        }
+
+        Ok(String::from_utf8_lossy(&output.stdout).to_string())
+    }
+
+    async fn git_status(&self, cwd: &Path, _args: serde_json::Value) -> anyhow::Result<ToolResult> {
+        let output = self
+            .run_git_command_in(cwd, &["status", "--porcelain=2", "--branch"])
+            .await?;
+
+        // Parse git status output into structured format
+        let mut result = serde_json::Map::new();
+        let mut branch = String::new();
+        let mut staged = Vec::new();
+        let mut unstaged = Vec::new();
+        let mut untracked = Vec::new();
+
+        for line in output.lines() {
+            if line.starts_with("# branch.head ") {
+                branch = line.trim_start_matches("# branch.head ").to_string();
+            } else if let Some(rest) = line.strip_prefix("1 ") {
+                // Ordinary changed entry
+                let mut parts = rest.splitn(3, ' ');
+                if let (Some(staging), Some(path)) = (parts.next(), parts.next()) {
+                    if !staging.is_empty() {
+                        let status_char = staging.chars().next().unwrap_or(' ');
+                        if status_char != '.' && status_char != ' ' {
+                            staged.push(json!({"path": path, "status": status_char}));
+                        }
+                        let status_char = staging.chars().nth(1).unwrap_or(' ');
+                        if status_char != '.' && status_char != ' ' {
+                            unstaged.push(json!({"path": path, "status": status_char}));
+                        }
+                    }
+                }
+            } else if let Some(rest) = line.strip_prefix("? ") {
+                untracked.push(rest.to_string());
+            }
+        }
+
+        result.insert("branch".to_string(), json!(branch));
+        result.insert("staged".to_string(), json!(staged));
+        result.insert("unstaged".to_string(), json!(unstaged));
+        result.insert("untracked".to_string(), json!(untracked));
+        result.insert(
+            "clean".to_string(),
+            json!(staged.is_empty() && unstaged.is_empty() && untracked.is_empty()),
+        );
+
+        let mut tr = ToolResult::success(serde_json::to_string_pretty(&result).unwrap_or_default());
+        tr.markdown_formatted = Some(render_status_markdown(&result));
+        Ok(tr)
+    }
+
+    async fn git_diff(&self, cwd: &Path, args: serde_json::Value) -> anyhow::Result<ToolResult> {
+        let files = args.get("files").and_then(|v| v.as_str()).unwrap_or(".");
+        let cached = args
+            .get("cached")
+            .and_then(|v| v.as_bool())
+            .unwrap_or(false);
+
+        // Validate files argument against injection patterns
+        self.sanitize_git_args(files)?;
+
+        let mut git_args = vec!["diff", "--unified=3"];
+        if cached {
+            git_args.push("--cached");
+        }
+        git_args.push("--");
+        git_args.push(files);
+
+        let output = self.run_git_command_in(cwd, &git_args).await?;
+
+        // Parse diff into structured hunks
+        let mut result = serde_json::Map::new();
+        let mut hunks = Vec::new();
+        let mut current_file = String::new();
+        let mut current_hunk = serde_json::Map::new();
+        let mut lines = Vec::new();
+
+        for line in output.lines() {
+            if line.starts_with("diff --git ") {
+                if !lines.is_empty() {
+                    current_hunk.insert("lines".to_string(), json!(lines));
+                    if !current_hunk.is_empty() {
+                        hunks.push(serde_json::Value::Object(current_hunk.clone()));
+                    }
+                    lines = Vec::new();
+                    current_hunk = serde_json::Map::new();
+                }
+                let parts: Vec<&str> = line.split_whitespace().collect();
+                if parts.len() >= 4 {
+                    current_file = parts[3].trim_start_matches("b/").to_string();
+                    current_hunk.insert("file".to_string(), json!(current_file));
+                }
+            } else if line.starts_with("@@ ") {
+                if !lines.is_empty() {
+                    current_hunk.insert("lines".to_string(), json!(lines));
+                    if !current_hunk.is_empty() {
+                        hunks.push(serde_json::Value::Object(current_hunk.clone()));
+                    }
+                    lines = Vec::new();
+                    current_hunk = serde_json::Map::new();
+                    current_hunk.insert("file".to_string(), json!(current_file));
+                }
+                current_hunk.insert("header".to_string(), json!(line));
+            } else if !line.is_empty() {
+                lines.push(json!({
+                    "text": line,
+                    "type": if line.starts_with('+') { "add" }
+                           else if line.starts_with('-') { "delete" }
+                           else { "context" }
+                }));
+            }
+        }
+
+        if !lines.is_empty() {
+            current_hunk.insert("lines".to_string(), json!(lines));
+            if !current_hunk.is_empty() {
+                hunks.push(serde_json::Value::Object(current_hunk));
+            }
+        }
+
+        result.insert("hunks".to_string(), json!(hunks));
+        result.insert("file_count".to_string(), json!(hunks.len()));
+
+        Ok(ToolResult::success(
+            serde_json::to_string_pretty(&result).unwrap_or_default(),
+        ))
+    }
+
+    async fn git_log(&self, cwd: &Path, args: serde_json::Value) -> anyhow::Result<ToolResult> {
+        let limit_raw = args.get("limit").and_then(|v| v.as_u64()).unwrap_or(10);
+        let limit = usize::try_from(limit_raw).unwrap_or(usize::MAX).min(1000);
+        let limit_str = limit.to_string();
+
+        let output = self
+            .run_git_command_in(
+                cwd,
+                &[
+                    "log",
+                    &format!("-{limit_str}"),
+                    "--pretty=format:%H|%an|%ae|%ad|%s",
+                    "--date=iso",
+                ],
+            )
+            .await?;
+
+        let mut commits = Vec::new();
+
+        for line in output.lines() {
+            let parts: Vec<&str> = line.split('|').collect();
+            if parts.len() >= 5 {
+                commits.push(json!({
+                    "hash": parts[0],
+                    "author": parts[1],
+                    "email": parts[2],
+                    "date": parts[3],
+                    "message": parts[4]
+                }));
+            }
+        }
+
+        let mut tr = ToolResult::success(
+            serde_json::to_string_pretty(&json!({ "commits": commits })).unwrap_or_default(),
+        );
+        tr.markdown_formatted = Some(render_log_markdown(&commits));
+        Ok(tr)
+    }
+
+    async fn git_branch(&self, cwd: &Path, _args: serde_json::Value) -> anyhow::Result<ToolResult> {
+        let output = self
+            .run_git_command_in(cwd, &["branch", "--format=%(refname:short)|%(HEAD)"])
+            .await?;
+
+        let mut branches = Vec::new();
+        let mut current = String::new();
+
+        for line in output.lines() {
+            if let Some((name, head)) = line.split_once('|') {
+                let is_current = head == "*";
+                if is_current {
+                    current = name.to_string();
+                }
+                branches.push(json!({
+                    "name": name,
+                    "current": is_current
+                }));
+            }
+        }
+
+        let mut tr = ToolResult::success(
+            serde_json::to_string_pretty(&json!({
+                "current": current,
+                "branches": branches
+            }))
+            .unwrap_or_default(),
+        );
+        tr.markdown_formatted = Some(render_branch_markdown(&current, &branches));
+        Ok(tr)
+    }
+
+    fn truncate_commit_message(message: &str) -> String {
+        if message.chars().count() > 2000 {
+            format!("{}...", message.chars().take(1997).collect::<String>())
+        } else {
+            message.to_string()
+        }
+    }
+
+    async fn git_commit(&self, cwd: &Path, args: serde_json::Value) -> anyhow::Result<ToolResult> {
+        let message = args
+            .get("message")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| anyhow::anyhow!("Missing 'message' parameter"))?;
+
+        // Sanitize commit message
+        let sanitized = message
+            .lines()
+            .map(|l| l.trim())
+            .filter(|l| !l.is_empty())
+            .collect::<Vec<_>>()
+            .join("\n");
+
+        if sanitized.is_empty() {
+            anyhow::bail!("Commit message cannot be empty");
+        }
+
+        // Limit message length
+        let message = Self::truncate_commit_message(&sanitized);
+
+        let output = self
+            .run_git_command_in(cwd, &["commit", "-m", &message])
+            .await;
+
+        match output {
+            Ok(_) => Ok(ToolResult::success(format!("Committed: {message}"))),
+            Err(e) => Ok(ToolResult::error(format!("Commit failed: {e}"))),
+        }
+    }
+
+    async fn git_add(&self, cwd: &Path, args: serde_json::Value) -> anyhow::Result<ToolResult> {
+        let paths = args
+            .get("paths")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| anyhow::anyhow!("Missing 'paths' parameter"))?;
+
+        // Validate paths against injection patterns
+        self.sanitize_git_args(paths)?;
+
+        let output = self.run_git_command_in(cwd, &["add", "--", paths]).await;
+
+        match output {
+            Ok(_) => Ok(ToolResult::success(format!("Staged: {paths}"))),
+            Err(e) => Ok(ToolResult::error(format!("Add failed: {e}"))),
+        }
+    }
+
+    async fn git_checkout(
+        &self,
+        cwd: &Path,
+        args: serde_json::Value,
+    ) -> anyhow::Result<ToolResult> {
+        let branch = args
+            .get("branch")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| anyhow::anyhow!("Missing 'branch' parameter"))?;
+
+        // Sanitize branch name
+        let sanitized = self.sanitize_git_args(branch)?;
+
+        if sanitized.is_empty() || sanitized.len() > 1 {
+            anyhow::bail!("Invalid branch specification");
+        }
+
+        let branch_name = &sanitized[0];
+
+        // Block dangerous branch names
+        if branch_name.contains('@') || branch_name.contains('^') || branch_name.contains('~') {
+            anyhow::bail!("Branch name contains invalid characters");
+        }
+
+        let output = self
+            .run_git_command_in(cwd, &["checkout", branch_name])
+            .await;
+
+        match output {
+            Ok(_) => Ok(ToolResult::success(format!(
+                "Switched to branch: {branch_name}"
+            ))),
+            Err(e) => Ok(ToolResult::error(format!("Checkout failed: {e}"))),
+        }
+    }
+
+    async fn git_stash(&self, cwd: &Path, args: serde_json::Value) -> anyhow::Result<ToolResult> {
+        let action = args
+            .get("action")
+            .and_then(|v| v.as_str())
+            .unwrap_or("push");
+
+        let output = match action {
+            "push" | "save" => {
+                self.run_git_command_in(cwd, &["stash", "push", "-m", "auto-stash"])
+                    .await
+            }
+            "pop" => self.run_git_command_in(cwd, &["stash", "pop"]).await,
+            "list" => self.run_git_command_in(cwd, &["stash", "list"]).await,
+            "drop" => {
+                let index_raw = args.get("index").and_then(|v| v.as_u64()).unwrap_or(0);
+                let index = i32::try_from(index_raw)
+                    .map_err(|_| anyhow::anyhow!("stash index too large: {index_raw}"))?;
+                self.run_git_command_in(cwd, &["stash", "drop", &format!("stash@{{{index}}}")])
+                    .await
+            }
+            _ => anyhow::bail!("Unknown stash action: {action}. Use: push, pop, list, drop"),
+        };
+
+        match output {
+            Ok(out) => Ok(ToolResult::success(out)),
+            Err(e) => Ok(ToolResult::error(format!("Stash {action} failed: {e}"))),
+        }
+    }
+}
